Updated: removed code that hides the sites filter
Removed the code that hides the items filter in the Select Sites UI element. Originally, it was designed to hide the filter when the select menu is not set by Site, but now the filter works with tags and clients to it is good to have it in all view.